Add ActionWriter base class
